Fix string reference comparison
The current implementation relies on the internal VM reference to
“android” be the same as the reference to the packageName and
processName objects. This particular if statement is not comparing the
actual value of the strings.

To reliably compare strings in Java with the equality operators every
non-statically internalized string must be suffixed with “.intern()”
Even that approach has some caveats.