FloatBuffer-related performance tuning
While Xposed maintains a cache of fields it’s still much faster to store
a reference to the field we need if we’re going to be calling it often
or need to call it quickly. No need to create a StringBuilder + String
and search for it in a HashMap each call to getObjectField.

This change also improves performance by only explicitly casting to
FloatBuffer once.

In less interesting news this change also more closely represents the
original ColorFade source.